requirements.txt
streamlit==1.31.0
easyocr==1.7.1
groq==0.4.2
huggingface-hub==0.20.3
python-dotenv==1.0.1
Pillow==10.2.0
numpy==1.24.3
httpx==0.27.2
transformers==4.40.2
torch==2.9.0
sentencepiece==0.2.1
